requirements.txt
streamlit
openai
langchain
langchain-community
langchain-huggingface
numpy
langchain-openai
faiss-cpu
sentence-transformers
PyPDF2
torch